ISSUE #155 - Attempt to fix single mesh generation

+++ bouncer/src/repo/core/model/bson/repo_bson_factory.cpp
@@ -132,11 +132,11 @@ LightNode RepoBSONFactory::makeLightNode(
 	const repo_color3d_t  &ambientColor,
 	const repo_color3d_t  &specularColor,
 	const repo_color3d_t  &diffuseColor,
-	const float           &innerConeAngle,
-	const float           &outerConeAngle,
-	const float           &attentuationConstant,
-	const float           &attentuationLinear,
-	const float           &attentuationQuadratic,
+	const double          &innerConeAngle,
+	const double          &outerConeAngle,
+	const double          &attentuationConstant,
+	const double          &attentuationLinear,
+	const double          &attentuationQuadratic,
 	const std::string     &name,
 	const int             &apiLevel)
 {
